Don't write the files section of xorg.conf by default.
If the user does want to write the files section, don't write out any font
paths, only write out the font server bit if they specify that.

This will depend on having xserver-xorg-core version 2:1.3.0.0.dfsg-3
which includes a patch to always look in the default font paths by
default. Bump the xserver-xorg dependency on xserver-xorg-core to this
version.