Slim down the modes we write out by default
The /config/display/modes now has no default (blank). As a result of this,
under normal conditions, this multi-select field will be empty.  The
debconfage and dexconf will not fill in this section at all. The server
will then choose the highest resolution that it can probe with DDC. This
should work for most people, who are probably choosing their best
resolution anyway. If xresprobe is available and used, the resolutions it
finds will be selected. Until there is a good way for the server to store
this information by itself, we should leave this as it is. Most people can
get by without this section, but it'll help some of those with broken
hardware.

If the user does select a mode or modes for this question, dexconf no
longer iterates through a set of depths and puts the mode list in for
each. Instead, if dexconf does find a set of specified modes, it will
write a single "Display" subsection for the screen, and it will contain
the modes. This is functionally equivalent to what we were doing anyway,
and it cuts down the number of lines in our xorg.conf considerably.

Conflicts:

	debian/changelog