Strip out use of glxinfo from apport hook
This tool is no longer provided in 'main' since the split-out of
mesa-utils (and it's not going to be re-added just for this).  Since we
already have alternate ways of gathering the info that actually use,
there isn't much value in conditionally using it if it's present and
omitting it if not, so just drop its use entirely.  For corner cases
where we do actually need the info, the user would probably need to be
told to install it anyway.