Use separate CSS & JS.
By default, asciidoc embeds the CSS and the JS (used to generate the
toc) in the generated HTML file. Tell it not to, through “-a linkcss”.
As a result, add the following files, from asciidoc 8.6.4:
 - xhtml11.css (as asciidoc-xhtml11.css)
 - asciidoc-xhtml11.js

and make xsf.css import it, with only a few additional tweaks.

Accordingly, perform substitutions in the generated HTML files to point
to the appropriate files (at the toplevel).